refactor: videoCardList roll back
