+++ src/js/components/SearchModal/SearchBar.js
@@ -2,6 +2,7 @@ import { getSearchAPI } from '../../api/api.js';
 import Component from '../../core/Component.js';
 import { rootStore } from '../../store/rootStore.js';
 import { webStore } from '../../store/WebStore.js';
+import request from '../../__mocks__/request.js';
 
 export default class SearchBar extends Component {
   template() {
@@ -37,7 +38,8 @@ export default class SearchBar extends Component {
 
       rootStore.setState({ isLoading: true });
 
-      const [error, data] = await getSearchAPI(query);
+      // const [error, data] = await getSearchAPI(query);
+      const [error, data] = await request();
       if (error) {
         rootStore.setState({
           isLoading: false,
